Fixed mini machines being usable on non-electronic recipes
Cackling Fiend mentioned that the intention with mini machines was for their use to be relegated to electronics recipes but this wasn't working with the current version. Categories need to be assigned during final fixes because mini machines for some reason overrides changes to their machines in their final fixes.